requirements.txt
streamlit
streamlit-tags
st-annotated-text